move fetch error handling to own module
