+++ packages/fetch/src/setup.ts
@@ -1,3 +1,4 @@
+import { newCustomError, throwRemoteError } from "./errors";
 import { format, parse } from "./parse";
 
 export function traverse(target: any, endpoint: string, path = "") {
@@ -15,7 +16,7 @@ export function traverse(target: any, endpoint: string, path = "") {
         return newCustomError(path);
 
       default:
-        throw new Error(`Unknown entity type ${target[0]}`)
+        throw new Error(`Unknown entity type ${target[0]}`);
     }
   }
 
@@ -28,43 +29,6 @@ export function traverse(target: any, endpoint: string, path = "") {
 
     return route;
   }
-}
-
-const CUSTOM_ERROR = new Map<string, typeof Error>();
-
-function newCustomError(path: string){
-  const match = /\/(\w+)$/.exec(path);
-  const uid = "/" + path.toLowerCase();
-
-  if(!match)
-    throw new Error("");
-
-  const ErrorType: typeof Error = new Function(`
-    return class ${match[1]} extends Error {}
-  `)();
-
-  CUSTOM_ERROR.set(uid, ErrorType);
-
-  return ErrorType;
-}
-
-function throwRemoteError(data: any){
-  const uid = data.error.toLowerCase();
-
-  const Type: typeof Error =
-    CUSTOM_ERROR.get(uid) || Error;
-
-  const error: any = new Type(data.message);
-
-  for(const key in data)
-    if(key == "stack"){
-      // const remoteLines = data.stack.map((x: string) => "    " + x);]
-      error.stack = error.stack.split("\n").splice(1, 2).join("\n");
-    }
-    else
-      (error as any)[key] = data[key];
-
-  return error;
 }
 
 function newHandler(path: string, endpoint: string){
